feat(chore): move useChatbotApi hook (graasp/graasp-app-code-capsule#161)
